Trying a new approach to the images
Method named displayTransitionImage_try2 does the same thing as
displayTransitionImage but in a different way. Rather than creating new
views each time a swipe or tap is recognized, it instead has an array
of views that is always transparent. When a gesture is recognized it
fades in and then out the view that is connected to that gesture.

I'm not sure which one is better. Like I'm not sure which one is faster
or which one is better in terms of memory management, just thought of
this way of doing it and thought I would write it up!